+++ web/src/popovers.js
@@ -2,6 +2,7 @@ import ClipboardJS from "clipboard";
 import {parseISO} from "date-fns";
 import $ from "jquery";
 import tippy, {hideAll} from "tippy.js";
+import url_template_lib from "url-template";
 
 import render_no_arrow_popover from "../templates/no_arrow_popover.hbs";
 import render_playground_links_popover_content from "../templates/playground_links_popover_content.hbs";
@@ -783,20 +784,21 @@ export function register_click_handlers() {
             const playground_info = realm_playground.get_playground_info_for_languages(
                 $codehilite_div.data("code-language"),
             );
-            // We do the code extraction here and set the target href combining the url_prefix
-            // and the extracted code. Depending on whether the language has multiple playground
-            // links configured, a popover is show.
+            // We do the code extraction here and set the target href expanding
+            // the url_template with the extracted code. Depending on whether
+            // the language has multiple playground links configured, a popover
+            // is shown.
             const extracted_code = $codehilite_div.find("code").text();
             if (playground_info.length === 1) {
-                const url_prefix = playground_info[0].url_prefix;
+                const url_template = url_template_lib.parse(playground_info[0].url_template);
                 $view_in_playground_button.attr(
                     "href",
-                    url_prefix + encodeURIComponent(extracted_code),
+                    url_template.expand({code: extracted_code}),
                 );
             } else {
                 for (const $playground of playground_info) {
-                    $playground.playground_url =
-                        $playground.url_prefix + encodeURIComponent(extracted_code);
+                    const url_template = url_template_lib.parse($playground.url_template);
+                    $playground.playground_url = url_template.expand({code: extracted_code});
                 }
                 toggle_playground_link_popover(this, playground_info);
             }