+++ zerver/openapi/python_examples.py
@@ -481,7 +481,7 @@ def add_realm_playground(client: Client) -> None:
     request = {
         "name": "Python playground",
         "pygments_language": "Python",
-        "url_prefix": "https://python.example.com",
+        "url_template": "https://python.example.com?code={code}",
     }
     result = client.call_endpoint(url="/realm/playgrounds", method="POST", request=request)
     # {code_example|end}